Update App.js and smallFunc.js
